[Add]: Color+Ext.swift
- MonthConfig.swift에서 Color 확장코드 분리
