+++ CalendarWidget/MonthConfig.swift
@@ -7,20 +7,6 @@
 
 import SwiftUI
 
-extension Color {
-    static let paleYellow   = Color(red: 252/255, green: 225/255, blue: 121/255)
-    static let palePink     = Color(red: 254/255, green: 138/255, blue: 138/255)
-    static let darkGreen    = Color(red: 0/255, green: 67/255, blue: 13/255)
-    static let paleGreen    = Color(red: 163/255, green: 230/255, blue: 127/255)
-    static let paleBlue     = Color(red: 139/255, green: 229/255, blue: 233/255)
-    static let skyBlue      = Color(red: 103/255, green: 155/255, blue: 197/255)
-    static let paleOrange   = Color(red: 197/255, green: 161/255, blue: 103/255)
-    static let darkOrange   = Color(red: 172/255, green: 110/255, blue: 16/255)
-    static let paleRed      = Color(red: 174/255, green: 80/255, blue: 80/255)
-    static let paleBrown    = Color(red: 124/255, green: 102/255, blue: 85/255)
-}
-
-
 struct MonthConfig {
     let backgroundColor: Color
     let emojiText: String
